Revert "Hiroki oikawa"
