+++ SkillTree/Assets/Scripts/InputText.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b9935b60cf8f9be4fbe8542d21bfda2a
+guid: 670387946480ea144a68b29de545d2a9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
